# Fix performance issue when RHS is a Matrix