Before this patch, `mul!` would incorrectly call `_generic_mapmat_mul!` instead of `_unsafe_mul!` in cases when the RHS was a matrix.

This problem lead to huge performance penalties in all cases where the LinearMap acts more efficiently on a matrix than on the individual columns.

```
using LinearMaps
using BenchmarkTools
using LinearAlgebra
A = randn(10000,1000);
AL = LinearMap(A);
B = randn(1000,1000);
C = zeros(size(A,1), size(B,2));
```

Before patch:
```
julia> @btime mul!($C,$A,$B);
  139.193 ms (0 allocations: 0 bytes)

julia> @btime mul!($C,$AL,$B);
  2.812 s (0 allocations: 0 bytes)
```

After patch:
```
julia> @btime mul!($C,$A,$B);
  142.395 ms (0 allocations: 0 bytes)

julia> @btime mul!($C,$AL,$B);
  147.012 ms (0 allocations: 0 bytes)
```