@dkarrasch 2021-03-04
Awesome! Nice catch. That logic was a bit tricky, and I see that we took the turn to the generic function one step too early. Could you bump the patch version number in Project.toml, then we merge and release. Thank you very much!