@rasmushenningsson 2021-03-04
Should be good to go! :)
Thanks for a nice package and a quick response!